remove "short" and "looong" stubs on header, and move pages to components directory
