+++ wwwroot/src/App.js
@@ -6,10 +6,10 @@ import {
 } from "react-router-dom";
 import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
 
-import SignIn from './pages/SignIn';
-import SignUp from './pages/SignUp';
-import Landing from './pages/Landing';
-import WhyUs from './pages/WhyUs';
+import SignIn from './components/SignIn';
+import SignUp from './components/SignUp';
+import Landing from './components/Landing';
+import WhyUs from './components/WhyUs';
 
 const theme = createMuiTheme({
   typography: {
@@ -81,8 +81,6 @@ function App() {
         <Route path="/signup"  element={<SignUp />} />
         <Route path="/signin"  element={<SignIn />} />
         <Route path="/why-us" element={<WhyUs />} />
-        <Route path="/short"  element={<WhyUs />} />
-        <Route path="/long"  element={<WhyUs />} />
     </Routes>
     </BrowserRouter>
   </ThemeProvider>
